Refactored ThreadCounts into its own file and added license headers to all files.
