+++ src/System.Private.CoreLib/src/System/Threading/ClrThreadPool.WorkerThread.Unix.cs
@@ -1,6 +1,8 @@
-﻿using System.Diagnostics;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
 using System.Globalization;
-using System.Runtime.InteropServices;
 using Internal.Runtime.Augments;
 
 namespace System.Threading
